BN-140 | Refactor. Rename routes.tsx as App.tsx
